+++ app/src/main/java/eu/faircode/email/MessageHelper.java
@@ -515,8 +515,8 @@ public class MessageHelper {
         return subject;
     }
 
-    Integer getSize() throws MessagingException {
-        int size = imessage.getSize();
+    Long getSize() throws MessagingException {
+        long size = imessage.getSize();
         if (size == 0)
             throw new MessagingException("Message empty");
         return (size < 0 ? null : size);
